+++ crace/utils/const.py
@@ -2,5 +2,7 @@ bold = '\033[1m'
 underline = '\033[4m'
 reset = '\033[0m'
 
+WIDTH = 100
+
 DELTA = 1e-4
 EPSILON = 1e-10